+++ src/zrb/task/llm_task.py
@@ -192,17 +192,19 @@ class LLMTask(BaseTask):
         self._max_call_iteration = max_call_iteration
         self._conversation_context = conversation_context
 
-    def add_tool(self, tool: ToolOrCallable):
-        self.append_tool(tool)
+    def add_tool(self, *tool: ToolOrCallable):
+        self.append_tool(*tool)
 
-    def append_tool(self, tool: ToolOrCallable):
-        self._additional_tools.append(tool)
+    def append_tool(self, *tool: ToolOrCallable):
+        for single_tool in tool:
+            self._additional_tools.append(single_tool)
 
-    def add_mcp_server(self, mcp_server: MCPServer):
-        self.append_mcp_server(mcp_server)
+    def add_mcp_server(self, *mcp_server: MCPServer):
+        self.append_mcp_server(*mcp_server)
 
-    def append_mcp_server(self, mcp_server: MCPServer):
-        self._additional_mcp_servers.append(mcp_server)
+    def append_mcp_server(self, *mcp_server: MCPServer):
+        for single_mcp_server in mcp_server:
+            self._additional_mcp_servers.append(single_mcp_server)
 
     def set_should_enrich_context(self, enrich_context: bool):
         self._should_enrich_context = enrich_context